requirements.txt
boto3>=1.28.53
boto3-stubs[boto3]
flywheel-gear-toolkit>=0.2
flywheel-sdk==16.19.2
fw-client>=0.7.0
fw-utils>=3
moto[s3,ssm]>=4.2.5
pandas>=2.1.1
pandas-stubs>=2.0.3
pyyaml==6.0
pytest>=7.2.0
pytz >= 2023.3
requests>=2.18.4
setuptools>=68.2.2
ssm-parameter-store>=19.11.0
types-PyYAML>=6.0.12
types-pytz >= 2023.3